--- a/go.work
+++ b/go.work
@@ -1,0 +1,3 @@
+go 1.19
+
+use ./hw10_program_optimization